Rename type_traits to policies
