+++ projects/rocprofiler-systems/source/lib/policies/thread_state_policy.hpp
@@ -8,16 +8,16 @@
 
 namespace rocprofsys
 {
-namespace type_traits
+namespace policies
 {
 // Structural concept, deliberately independent of any concrete state type
 // (e.g. rocprofsys::state::thread) so that lower-level code depending on this
 // header never needs to include a higher-level state module.
 template <typename T>
-concept thread_state_policy = requires {
+concept thread_state_policy = requires(typename T::State state) {
     typename T::State;
     { T::Internal } -> std::convertible_to<typename T::State>;
-    { T::scoped(std::declval<typename T::State>()) } -> std::destructible;
+    { T::scoped(state) } -> std::destructible;
 };
-}  // namespace type_traits
+}  // namespace policies
 }  // namespace rocprofsys
